refactor: move iroh bytes get handlers out of node into module
