+++ iroh/src/lib.rs
@@ -8,6 +8,7 @@ pub mod baomap;
 #[cfg(feature = "iroh-collection")]
 pub mod collection;
 pub mod dial;
+pub mod get;
 pub mod node;
 pub mod rpc_protocol;
 pub mod util;
